+++ app/admin/internal/data/ent/userdepartment.go
@@ -18,7 +18,7 @@ import (
 type UserDepartment struct {
 	config `json:"-"`
 	// ID of the ent.
-	ID int64 `json:"id,omitempty"`
+	ID int `json:"id,omitempty"`
 	// SysUser ID
 	UserID int64 `json:"user_id,omitempty"`
 	// Department ID
@@ -95,7 +95,7 @@ func (ud *UserDepartment) assignValues(columns []string, values []any) error {
 			if !ok {
 				return fmt.Errorf("unexpected type %T for field id", value)
 			}
-			ud.ID = int64(value.Int64)
+			ud.ID = int(value.Int64)
 		case userdepartment.FieldUserID:
 			if value, ok := values[i].(*sql.NullInt64); !ok {
 				return fmt.Errorf("unexpected type %T for field user_id", values[i])
