+++ app/admin/internal/data/ent/usertenant/where.go
@@ -9,47 +9,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int64) predicate.UserTenant {
+func ID(id int) predicate.UserTenant {
 	return predicate.UserTenant(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int64) predicate.UserTenant {
+func IDEQ(id int) predicate.UserTenant {
 	return predicate.UserTenant(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int64) predicate.UserTenant {
+func IDNEQ(id int) predicate.UserTenant {
 	return predicate.UserTenant(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int64) predicate.UserTenant {
+func IDIn(ids ...int) predicate.UserTenant {
 	return predicate.UserTenant(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int64) predicate.UserTenant {
+func IDNotIn(ids ...int) predicate.UserTenant {
 	return predicate.UserTenant(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int64) predicate.UserTenant {
+func IDGT(id int) predicate.UserTenant {
 	return predicate.UserTenant(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int64) predicate.UserTenant {
+func IDGTE(id int) predicate.UserTenant {
 	return predicate.UserTenant(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int64) predicate.UserTenant {
+func IDLT(id int) predicate.UserTenant {
 	return predicate.UserTenant(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int64) predicate.UserTenant {
+func IDLTE(id int) predicate.UserTenant {
 	return predicate.UserTenant(sql.FieldLTE(FieldID, id))
 }
